requirements.txt
livekit-agents
openai
httpx
python-dotenv
